+++ arcangelo/arcangelo/tests/data/match/odoo7.py
@@ -7,7 +7,7 @@ class ResPartner(orm.osv):
     _inherit = "res.partner"
 
     def unlink(self, cr, uid, param, context=None):
-        return super(ResPartner, self).unlink(cr, uid, param, context=context)
+        return super(ResPartner, self).unlink(cr, uid, param)
 
     def compute(self, cr, uid, param, context=None):
-        return super(ResPartner, self).unlink(cr, uid, param, context=context)
+        return super(ResPartner, self).unlink(cr, uid, param)