+++ arcangelo/arcangelo/tests/data/match/openerp7_old_api.py
@@ -4,7 +4,12 @@ from openerp.osv import fields, osv
 from openerp.osv.orm import except_orm
 
 
-class ResExample(osv.Model):
+class ResPartner(osv.Model):
+    _inherit = "res.partner"
+
+    def unlink(self, cr, uid, param, context=None):
+        return super(ResPartner, self).unlink(cr, uid, param)
+
     def my_fun(self, cr, uid, param, context=None):
-        super(ResExample, self).my_fun(cr, uid, param, context=context)
+        super(ResPartner, self).my_fun(cr, uid, param)
         raise except_orm("Error", "Example")