+++ arcangelo/arcangelo/tests/data/pkg_py3_02.py
@@ -1,9 +1,8 @@
 #! /usr/bin/env python
 # flake8: noqa - pylint: skip-file
-# -*- coding: utf-8 -*-
 
 
-class ResExample(object):
+class MyExample(object):
     def __init__(self):
         pass