+++ arcangelo/arcangelo/egg-info/CHANGELOG.rst
@@ -1,10 +1,19 @@
+2.1.1 (2025-06-28)
+~~~~~~~~~~~~~~~~~~~
+
+* [IMP] New trigger search rather than match in rules
+* [IMP] Two passes parsing
+* [IMP] New pass1 context
+* [IMP] Set trigger with parameters
+* [FIX] New rule parsing algorithm
+
 2.1.0 (2025-06-15)
 ~~~~~~~~~~~~~~~~~~~
 
 * [IMP] Split from wok_code
 * [IMP] Graphical files are copied only if they does not exist on target
 * [IMP] Before migration warns on different base name
-* [FIX] If target direcotry does not exist, will be create
+* [FIX] If target directory does not exist, will be create
 
 2.0.22 (2025-05-31)
 ~~~~~~~~~~~~~~~~~~~